🚚 [packages] Move class `DefaultPackageRepository` to module `repository`
